7-network_dedicated_container/pyproject.toml
[project]
name = "7-network-dedicated-container"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mysql-connector-python>=8.0.0",
    "pytest>=8.0.0",
    "pytest-timeout>=2.0.0",
]
